order-processing/requirements.txt
Django==1.8.7
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
XlsxWriter==0.7.7
boto==2.38.0
chardet==2.0.1
defer==1.0.6
django-admin-bootstrapped==2.5.6
django-import-export==0.2.9
django-suit==0.2.15
paramiko==1.16.0
xlrd==0.9.4
xlwt==1.0.0
